+++ main.py
@@ -20,7 +20,11 @@ app.include_router(
 )
 
 app.include_router(
-    schemas_from_user.routes.user_v1, 
+    schemas_from_user.routes.user_info_v1, 
     prefix=f"{schemas_from_user.path_segment}{schemas_from_user.segments_path.info}"
 )
 
+app.include_router(
+    schemas_from_user.routes.user_sync_v1, 
+    prefix=f"{schemas_from_user.path_segment}{schemas_from_user.segments_path.sync}"
+)