+++ src/presentation/routes/user_route.py
@@ -1,16 +1,18 @@
 from fastapi import APIRouter, status
 from fastapi.params import Depends
 
-from src.presentation.handler.user_handler import get_all_user_info_from_db
+from src.domain.models.user_model import FirebaseUserModel
+from src.presentation.handler.user_handler import get_all_user_info_from_db, sign_up_user, log_in_user
 from ..handler.responses import MyResponse
 
 from src.presentation.handler.auth import verify_api_key
 
 
 
-user_v1 = APIRouter(prefix='/v1')
+user_info_v1 = APIRouter(prefix='/v1')
+user_sync_v1 = APIRouter(prefix='/v1')
 
-@user_v1.get(f"/all",
+@user_info_v1.get(f"/all",
             summary="Get All User Info", tags=["User"],
             description="This endpoint returns all user information.",
             response_model=MyResponse,
@@ -23,4 +25,41 @@ def get_all_user_info(api_key_secret: str = Depends(verify_api_key)):
         status=status.HTTP_200_OK,
         message="All user information retrieved successfully.",
         data=all_user_info
+    )
+
+@user_sync_v1.post(f"/sign-up",
+             summary="Sign Up User", tags=["User"],
+             description="This endpoint handles the creation of a new user in both the database and Firebase.",
+             response_model=MyResponse,
+             status_code=status.HTTP_201_CREATED,
+
+             )
+def post_sign_up_user(raw_user_data: FirebaseUserModel, api_key_secret: str = Depends(verify_api_key)):
+    """
+    Sign up a new user.
+    This endpoint handles the creation of a new user in both the database and Firebase.
+    """
+
+
+    return MyResponse(
+        status=status.HTTP_201_CREATED,
+        message="User signed up successfully.",
+        data=sign_up_user(raw_user_data)
+    )
+
+@user_sync_v1.patch(f"/log-in",
+             summary="Log In User", tags=["User"],
+             description="This endpoint handles user authentication and returns user data if successful.",
+             response_model=MyResponse,
+             status_code=status.HTTP_200_OK)
+def post_log_in_user(raw_user_data: FirebaseUserModel, api_key_secret: str = Depends(verify_api_key)):
+    """
+    Log in a user.
+    This endpoint handles user authentication and returns user data if successful.
+    """
+
+    return MyResponse(
+        status=status.HTTP_200_OK,
+        message="User logged in successfully.",
+        data=log_in_user(raw_user_data)
     )