+++ src/presentation/schema/user_schema.py
@@ -4,7 +4,7 @@ from abc import abstractmethod
 from dataclasses import dataclass
 from fastapi import APIRouter
 from src.presentation.routes.hello_user_route import hello_user_v1
-from src.presentation.routes.user_route import user_v1
+from src.presentation.routes.user_route import user_info_v1, user_sync_v1
 
 @abstractmethod
 @dataclass
@@ -27,7 +27,9 @@ class UserBaseSchema:
 class UserRoutes:
 
     hello_user_v1 = hello_user_v1
-    user_v1 = user_v1
+    user_info_v1 = user_info_v1
+    user_sync_v1 = user_sync_v1
+
 
 
 class UserSegmentsPath:
@@ -35,6 +37,7 @@ class UserSegmentsPath:
     Available segments for user-related endpoints.
     """
     info = "/info"
+    sync = "/sync"
 
 
 class UserSchema(UserBaseSchema):